commerce_flask_python/requirements.txt
auth0-python==3.24.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
click==8.0.3
cryptography==35.0.0
dnspython==2.2.1
Flask==2.0.2
Flask-Cors==3.0.10
flask-talisman==0.8.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
PyJWT==2.3.0
pymongo==4.3.3
python-dotenv==0.19.1
requests==2.28.1
six==1.16.0
urllib3==1.26.13
Werkzeug==2.0.2
